+++ ChangeLog.md
@@ -42,3 +42,4 @@
 ## 0.3.0.0
 
 * Removed `Show` instance for `Result`.  `Result` output should now be acquired with `resultToString`.
+* Renamed `isPassed` to `resultIsPassed` for consistency with `resultToString`.